requirements.txt
dlib==19.24.2
face-recognition==1.3.0
streamlit==1.31.1
opencv-python==4.9.0.80
pandas==2.2.1
numpy==1.26.4
Pillow==10.2.0
sqlalchemy==2.0.27
datetime 